remove lib/app.js, merge with lib/entrypoint
