+++ src/lib/pages/newsletter.js
@@ -6,4 +6,4 @@
 import '../components/Subscribe';
 import '../components/Newsletter';
 
-import '../app';
+import '../components/base';
